external/network-device-mcp-server/backup_20250930_141530/src/integrations/fortianalyzer.py
"""
FortiAnalyzer Integration - Production Version
Manages FortiAnalyzer log collection, analysis, and reporting functionality
NO MOCK DATA - Returns empty results when FortiAnalyzer is not configured
"""

import os
import sys
import json
from datetime import datetime, timedelta
from typing import Dict, List, Optional, Any
import urllib3

# Disable SSL warnings for FortiAnalyzer connections
urllib3.disable_warnings(urllib3.exceptions.InsecureRequestWarning)

class FortiAnalyzerManager:
    """
    Manages FortiAnalyzer operations for log analysis and security intelligence
    Production version - no mock data
    """
    
    def __init__(self, config_path: str = None):
        """
        Initialize FortiAnalyzer Manager
        
        Args:
            config_path: Path to FortiAnalyzer configuration
        """
        self.config_path = config_path or os.path.join(os.path.dirname(__file__), '..', 'config.py')
        self.sessions = {}  # Track active FortiAnalyzer sessions
        self.configured = False  # Track if FortiAnalyzer is properly configured
        
    def get_fortianalyzer_instances(self) -> Dict[str, Any]:
        """
        Get configured FortiAnalyzer instances
        
        Returns:
            Dictionary containing FortiAnalyzer instances and their configurations
        """
        return {
            "success": True,
            "fortianalyzer_instances": [],
            "message": "FortiAnalyzer integration ready - requires configuration",
            "configuration_required": [
                "FortiAnalyzer host/IP address",
                "Admin username and password", 
                "API access permissions",
                "Device group configurations"
            ]
        }
    
    def get_security_logs(self, brand: str, store_id: str, timeframe: str = "1h", log_type: str = "traffic") -> Dict[str, Any]:
        """
        Get security logs for a specific store
        
        Args:
            brand: Restaurant brand (BWW, ARBYS, SONIC)
            store_id: Store identifier
            timeframe: Time period for logs (1h, 24h, 7d, 30d)
            log_type: Type of logs (traffic, utm, event, dns)
            
        Returns:
            Dictionary containing log analysis results
        """
        if not self.configured:
            return {
                "success": False,
                "error": "FortiAnalyzer not configured",
                "message": "FortiAnalyzer integration requires proper configuration to access real log data",
                "logs": [],
                "analysis": {
                    "total_events": 0,
                    "security_events": 0,
                    "blocked_attempts": 0,
                    "allowed_traffic": 0
                }
            }
        
        # TODO: Implement actual FortiAnalyzer API calls
        return {
            "success": True,
            "logs": [],
            "analysis": {
                "total_events": 0,
                "security_events": 0,
                "blocked_attempts": 0,
                "allowed_traffic": 0,
                "message": "No log data available for the specified timeframe"
            }
        }
    
    def get_threat_intelligence(self, brand: str, timeframe: str = "24h") -> Dict[str, Any]:
        """
        Get threat intelligence for specific brand
        
        Args:
            brand: Restaurant brand
            timeframe: Time period for analysis
            
        Returns:
            Dictionary containing threat intelligence
        """
        return {
            "success": True,
            "threat_intelligence": {
                "total_threats": 0,
                "threat_categories": {},
                "top_sources": [],
                "blocked_ips": [],
                "malware_detected": 0,
                "intrusion_attempts": 0
            },
            "message": "Threat intelligence requires FortiAnalyzer configuration"
        }
    
    def get_brand_analytics(self, brand: str, metric_type: str = "security") -> Dict[str, Any]:
        """
        Get analytics for specific brand
        
        Args:
            brand: Restaurant brand
            metric_type: Type of metrics (security, performance, bandwidth)
            
        Returns:
            Dictionary containing analytics data
        """
        return {
            "success": True,
            "analytics": {
                "bandwidth_usage": {"total_gb": 0, "peak_hours": []},
                "security_events": {"total": 0, "categories": {}},
                "performance_metrics": {"avg_latency_ms": 0, "packet_loss_percent": 0}
            },
            "message": "Analytics require FortiAnalyzer configuration and historical data"
        }
    
    def search_logs(self, query: str, timeframe: str = "24h", brands: List[str] = None) -> Dict[str, Any]:
        """
        Search logs across FortiAnalyzer instances
        
        Args:
            query: Search query string
            timeframe: Time period for search
            brands: List of brands to search (optional)
            
        Returns:
            Dictionary containing search results
        """
        return {
            "success": True,
            "search_results": [],
            "query": query,
            "timeframe": timeframe,
            "total_matches": 0,
            "message": "Log search requires FortiAnalyzer configuration"
        }
    
    def generate_security_report(self, brand: str, timeframe: str = "7d") -> Dict[str, Any]:
        """
        Generate comprehensive security report for brand
        
        Args:
            brand: Restaurant brand
            timeframe: Report time period
            
        Returns:
            Dictionary containing report data
        """
        return {
            "success": True,
            "report": {
                "brand": brand,
                "timeframe": timeframe,
                "generated_at": datetime.now().isoformat(),
                "summary": {
                    "total_events": 0,
                    "security_incidents": 0,
                    "policy_violations": 0,
                    "system_alerts": 0
                },
                "recommendations": [
                    "Configure FortiAnalyzer integration for detailed security reporting",
                    "Set up log forwarding from FortiGate devices",
                    "Enable automated threat detection rules"
                ]
            },
            "message": "Security reports require FortiAnalyzer configuration"
        }